requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-environ==0.4.5
drf-spectacular==0.17.3
model-bakery==1.3.2
parameterized==0.8.1
gunicorn==20.1.0
psycopg2-binary==2.9.1